// src/app/core/enums/discard-reason.enum.ts
export enum DiscardReason {
  NotAFeature = 'NotAFeature',
  GeometryNotPoint = 'GeometryNotPoint',
  InvalidCoordinates = 'InvalidCoordinates',
  CoordinatesOutOfRange = 'CoordinatesOutOfRange',
  MissingProperties = 'MissingProperties',
  InvalidName = 'InvalidName',
  InvalidCategory = 'InvalidCategory',
}